# src/main/java/com/example/authservice/exception/ErrorResponse.java
package com.example.authservice.exception;


import lombok.AllArgsConstructor;
import lombok.Getter;

@Getter
@AllArgsConstructor
public class ErrorResponse {
    private String code;
    private String message;
}