# src/main/java/com/example/authservice/request/RefreshTokenRequest.java
package com.example.authservice.request;

import jakarta.validation.constraints.NotBlank;

public record RefreshTokenRequest(
        @NotBlank(message = "Refresh token is required") String refreshToken) {
}